Telegram-bot/requirements.txt
# This file contains the dependencies for the Telegram bot project.
python-telegram-bot==13.15
python-dotenv
gspread
oauth2client
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib